economy and player hp has been set up, wavemanager and gamemanager has been rework, all logical process and UI changes are done in the game manager, wavemanager only executes, start waves and instaciate enemies, flying enemies have been set up logic for enemy movement has been set up to function with flying and ground enemies wave manager now supports multiple spawn points, is not fully functional but the arquitecture is there in terms of implementation in the future. all enemy info has been place in SOs, no data is held by prefabs or scripts, those only handle logic economy is set up as shown in the documentation the game currently has not been set up, the next push will be after the set up is done and the set up is completed.

+++ Assets/Scripts/ScriptableObjects/Enemies/EnemyInfo.cs
@@ -5,9 +5,9 @@ public class EnemyInfo : ScriptableObject
     public string enemyName;
     public GameObject enemyPrefab;
 
-    public float maxHealth;
+    public int maxHealth;
     public float moveSpeed;
-    public float dangerLevel;
+    public int dangerLevel;
 
 
     public bool isAerial;
